pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loci-extract"
version = "0.1.0"
description = "Standalone document extraction and OCR service for ephemeral ingestion pipelines"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "sudobot99" }]
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "pymupdf>=1.24,<2.0",
  "python-docx>=1.1,<2.0",
  "openpyxl>=3.1,<4.0",
  "pandas>=2.2,<3.0",
  "pillow>=10.4,<11.0",
  "pytesseract>=0.3.10,<1.0",
  "pdfplumber>=0.11,<1.0",
  "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "httpx>=0.27,<1.0",
  "ruff>=0.6,<1.0",
]
gpu = [
  "paddlepaddle-gpu>=2.6,<3.0",
  "paddleocr>=2.7,<3.0",
  "numpy>=1.24,<3.0",
  "transformers>=4.40,<5.0",
  "torch>=2.2,<3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
